refactor : keep component and container in the same file
